flask_app/requirements.txt
Flask
Flask-Limiter
Flask-WTF
Werkzeug
python-dotenv